aligned footer-left
The reason I prefer this approach is that there are fewer values we have to maintain. If we decide to change the spacing, the elements will have 0 margin and therefore will still be aligned. If we instead have a 7% margin for one element, a 10% margin for another and a 10px one for another, once the parent changes we'll have to readjust the values because the child elements don't change at the same rate