Merged changes of the other pr
With the changes of the footer, the horizontal scrolling is gone in Home, About and Project pages. I'm not completely faithful to the design in those 2 places:
(1) The ubc logo in the footer is bigger in this commit because the text would have been unreadable otherwise.
(2) The horizontal line in the footer is aligned with the rest of the content to make it look neat.
Another comment: footer-right is currently not horizontally aligned to anything as per the design files. The more I look at it the more I want to change it. What do you think?